Fix books routes

+++ server/routes/bookcases.js
@@ -1,12 +1,15 @@
 const express = require('express');
 const asyncHandler = require('../middleware/async');
 const Bookcase = require('../models/Bookcase');
+const Book = require('../models/Book');
 const Swap = require('../models/Swap');
 const router = express.Router();
 const { protect, authorize } = require('../middleware/auth');
 
 router.get(
     '/',
+    protect,
+    authorize('admin'),
     asyncHandler(async (req, res, next) => {
         const bookcases = await Bookcase.find()
             .populate({
@@ -43,35 +46,40 @@ router.get(
 );
 
 router.get(
-    '/:id',
+    '/me',
+    protect,
     asyncHandler(async (req, res, next) => {
         try {
-            const bookcase = await Bookcase.findById(req.params.id)
-                .populate({
-                    path: 'parentBook',
-                    select: '-title -_id',
-                    populate: [
-                        {
-                            path: 'author',
-                            select: 'name -_id',
-                        },
-                        {
-                            path: 'category',
-                            select: 'name -_id',
-                        },
-                        {
-                            path: 'publisher',
-                            select: 'name -_id',
-                        },
-                    ],
-                })
-                .populate({
-                    path: 'owner',
-                    select: 'name email -_id',
-                })
-                .populate({
-                    path: 'swaps',
-                });
+            // const bookcase = await Bookcase.findById(req.params.id)
+            const bookcase = await Bookcase.find(
+                {
+                    owner: req.user.id,
+                },
+                {
+                    change: 1,
+                    title: 1,
+                }
+            ).populate({
+                path: 'parentBook',
+                select: '-title -_id',
+                populate: [
+                    {
+                        path: 'author',
+                        select: 'name -_id',
+                    },
+                    {
+                        path: 'category',
+                        select: 'name -_id',
+                    },
+                    {
+                        path: 'publisher',
+                        select: 'name -_id',
+                    },
+                ],
+            })
+            .populate({
+                path: 'swaps',
+            });
 
             res.status(200).json({
                 success: true,
@@ -87,10 +95,39 @@ router.post(
     '/',
     protect,
     asyncHandler(async (req, res, next) => {
+        let bookcase;
+        console.log('post bookcase...');
         const { error } = Bookcase.validateBookcase(req.body);
-        if (error) return res.status(400).send(error.details[0].message);
 
-        const bookcase = await Bookcase.create(req.body);
+        if (error) {
+            return res.status(400).send(error);
+        }
+
+        let parentBook = await Book.findOne({
+            title: req.body.title,
+        });
+
+        if (!parentBook) {
+            console.log('Create parent book first....');
+            const error_book = Book.validateBook(req.body);
+            if (error_book) {
+                return res.status(400).send("Title, author, publisher and category required");
+            }
+
+            parentBook = await Book.create({
+                title: req.body.title,
+                author: req.body.author,
+                publisher: req.body.publisher,
+                category: req.body.category,
+            });
+        }
+        console.log('Create bookcase....');
+        bookcase = await Bookcase.create({
+            owner: req.user.id,
+            change: req.body.change,
+            title: req.body.title,
+            parentBook: parentBook,
+        });
 
         res.status(201).json({
             success: true,
@@ -98,6 +135,7 @@ router.post(
         });
     })
 );
+
 
 // @desc    Swap book
 // @route   POST /api/bookcases/:id/swaps
@@ -138,15 +176,24 @@ router.put(
     protect,
     asyncHandler(async (req, res, next) => {
         try {
-            const bookcase = await Bookcase.findByIdAndUpdate(req.params.id, req.body, {
-                new: true,
-                runValidators: true,
-            });
-
-            res.status(200).json({
-                success: true,
-                data: bookcase,
-            });
+            const bookcase = await Bookcase.findById(req.params.id);
+            console.log(bookcase)
+            if(bookcase.owner==req.user.id) {
+                bookcase.owner=req.user.id;
+                bookcase.change=req.body.change
+                const bookcase_update = await Bookcase.findByIdAndUpdate(req.params.id, bookcase, {
+                    new: true,
+                    runValidators: true,
+                });
+                res.status(200).json({
+                    success: true,
+                    data: bookcase_update,
+                });
+            }
+            else {
+                res.status(401).send("You don't have permission");
+            }
+            
         } catch {
             res.status(404).send('The book with the given id was not found.');
         }
@@ -158,14 +205,19 @@ router.delete(
     protect,
     asyncHandler(async (req, res, next) => {
         try {
-            const result = await Bookcase.deleteOne({
-                _id: req.params.id,
-            });
-
-            res.status(200).json({
-                success: true,
-                data: result,
-            });
+            const bookcase = await Bookcase.findById(req.params.id);
+            console.log(bookcase)
+            if (bookcase.owner==req.user.id) {
+                const result = await Bookcase.deleteOne({
+                    _id: req.params.id,
+                });
+                res.status(200).json({
+                    success: true,
+                });
+            }
+            else {
+                return res.status(401).send("You don't have permission");
+            }
         } catch {
             res.status(404).send('The book with the given id was not found.');
         }
